move guest model class into a model package in guest module
