+++ app/guest/__init__.py
@@ -2,6 +2,5 @@
 from flask import Blueprint
 from . import add_guest
 
-
 blueprint = Blueprint('guest', __name__)
 add_guest.routes(blueprint)
