+++ app/model/__init__.py
@@ -1,3 +1,2 @@
 # flake8:noqa
 from app.model.invite import Invite
-from app.model.guest import Guest
